when u were here before couldnt look you in the eyes

+++ pushswap/ft_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: dinda-si <dinda-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 16:45:21 by dinda-si          #+#    #+#             */
-/*   Updated: 2024/03/05 17:32:21 by dinda-si         ###   ########.fr       */
+/*   Updated: 2024/03/12 17:50:50 by dinda-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,9 @@
 # include <stdlib.h>
 # include <string.h>
 # include <unistd.h>
+# include <limits.h>
 
-int					ft_atoi(const char *str);
+long int			ft_atoi(const char *str);
 void				*ft_calloc(size_t nmemb, size_t size);
 char				*ft_itoa(int n);
 int					ft_putchar_fd(char c, int fd);
